backend/requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
authlib
itsdangerous
pydantic